backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
boto3==1.34.39
pydantic==2.5.0
pydantic-settings==2.1.0